A little problem in the implementation of the type system has led to a number of big changes. The problem was that 'void' can basically represent any type, since it is the bottom type of the hierarchy and is a subtype of all types. In the previous implementation void inherited directly from the abstract Type class, which knew nothing about all other types. This made it impossible for void to represent other types and eventually led to many case distinctions in client code which needed to compute the difference between empty sets, empty relations, etc.
The void type comes into play especially often when dealing with empty containers, [] is a list[void], {} is a set[void], etc.

To fix the issue I have pulled up all methods of specific types to the abstract Type. Specific types override these if they have something sensible to do. With this design, suddenly the case distinctions for NamedType can also dissappear since (like void) they may be subtypes of any other types. By overriding all methods, NamedType can now simulate to be the type it derived from. The getBaseType() method has therefore dissappeared along with all of its calls and case distinctions in the client code.

As a side-effect it was also need to change the semantics of the getType() method on INode, and to remove getTreeNodeType() from INode(). getType() now simply returns the TreeNodeType, from which the original NamedTreeType could be retrieved if necessary.

Needless to say these changes have had a huge impact, both internally and in any client code of the PDB.

All tests run successfully.